Bump precision

+++ test/rulesets/Base/array.jl
@@ -1,7 +1,7 @@
 @testset "constructors" begin
     @testset "undef" begin
         test_rrule(Array{Float64, 1}, undef, 5)
-        test_rrule(Array{Float32, 3}, undef, 5, 4, 3)
+        test_rrule(Array{Float64, 3}, undef, 5, 4, 3)
     end
     @testset "from existing array" begin
         test_rrule(Array, randn(2, 5))
